Add error-handling and updated documentation.
Added doc to switch mibtool.py over to pysmi library which includes it.

Added some error checking to tell if pysmi is not installed, and exit function. Function appears to complete successfully without it unless you look higher in stdout and notice the errors. Beautifulsoup fails on script launch since it has a defined import - pysmi (mibtool.py) does not, this method makes it easier to tell at the beginning that there are additional libraries required.